% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorstuff.R
\name{bb_cols}
\alias{bb_cols}
\title{Function to extract bb colors as hex codes}
\usage{
bb_cols(...)
}
\arguments{
\item{...}{Character names of default_colors}
}
\value{
vector of colors (Hex-format)
}
\description{
Function to extract bb colors as hex codes
}
\examples{
colors<-bb_cols("Bermuda Gray","Downy")

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}
