% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorstuff.R
\name{bb_pal}
\alias{bb_pal}
\title{Return function to interpolate a bb color palette}
\usage{
bb_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in drsimonj_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\value{
vector of colors (Hex-format)
}
\description{
Return function to interpolate a bb color palette
}
\examples{
colors<-bb_cols("Bermuda Gray","Downy")

}
\author{
Stefan Bartels, \email{email@biobits.eu}
}
