% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhelper.R
\name{getDBConnect}
\alias{getDBConnect}
\title{R Function für generating a custom colorpallette of length x}
\usage{
getDBConnect(dbase = "ucchscience", dbuser = NULL, dbpwd = NULL,
  useconfig = T, configfile = NULL)
}
\arguments{
\item{dbase}{:character vector - Name of ODBC-Connection as specified in odbc.ini}

\item{dbuser}{:character vector - Username for database}

\item{dbpwd}{:character vector - DB Password}

\item{useconfig}{:Boolean - If True the functions tries to read username and password from a config YAML File (default: "/etc/odbc_security.yml")}

\item{configfile}{:character vector - Path to config-YAML file with stored credentials}
}
\value{
an odbc database connection
}
\description{
R Function für generating a custom colorpallette of length x
}
\examples{
con<-getDBConnect(dbase= "mydatabase")

con <-getDBConnect("mydatabase",dbuser="User",dbpwd="password",useconfig=F)



}
\author{
Stefan Bartels, \email{email@biobits.eu}
}
