% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorstuff.R
\name{theme_bb}
\alias{theme_bb}
\title{Fill scale constructor for bb colors}
\usage{
theme_bb(theme = "main", ...)
}
\arguments{
\item{theme}{Character name of theme in bb_palettes}

\item{...}{Additional arguments passed to theme()}
}
\description{
Fill scale constructor for bb colors
}
